refactor: re-point v1 contract tests to v1 errors module
